Add default values to NbnQuery model properties
The locationID and collector properties in the NbnQuery model are
expected in the views. If the data returned from the NBN API does not
include these fields, then the view will fail with an error; e.g.,

    Undefined property: stdClass::$locationId

This commit adds default values to the model in cases where those
required fields/properties are missing. It uses the new PHP 7 null
coalescing operator (https://roper.im/null-coalescing).