Remove tabs on large and above breakpoints
On large and above breakpoints (>= 992px) we need to do several things:

- Save the currently selected tab
- Hide the tab bar (#tabs) at the top
- Force both of the tab content areas to show side by side

I'm doing this by removing the Bootstrap tab-content class on the #tab-content
container and adding the show class to both tab content areas.

When we move back to a small breakpoint, we need to restore the tabs as they
were by restoring the tab-content class and using the Bootstrap API to show the
active tab.

Also, to support progressive enhancement, I am only adding the tab-content class
on the page load event. This, in effect, disables the tabs if JS doesn't
load and forces the data to always show (but the map won't load).

I'm not that happy about the flash of data below the map on small
screens as the page loads, but can revisit that later.