Highlight correponding data rows
When clicking or tapping on a record, a popup containing Site, Grid,
and Recorder(s) is shown *and* the corresponding data row is
highlighted.

When the popup is dismissed, the row highlight is removed.

Doing this is helpful in understanding what the red dots represent.
It is essential for orientation if using a small screen and tabs.